requirements.txt
python-novaclient
pyyaml
